EAI functions moved to eai.c
