+++ Makefile
@@ -1593,7 +1593,6 @@ env.a control.o constmap.o str.a fs.a auto_qmail.o auto_split.o auto_uids.o
         date822fmt.o wildmat.o qregex.o env.a control.o constmap.o datetime.a case.a seek.a ndelay.a open.a sig.a \
         getln.a stralloc.a alloc.a substdio.a error.a str.a fs.a auto_qmail.o auto_split.o auto_uids.o
 
-
 qmail-queue.0: \
 qmail-queue.8
 	nroff -man qmail-queue.8 > qmail-queue.0
@@ -1604,19 +1603,23 @@ alloc.h substdio.h datetime.h now.h datetime.h triggerpull.h extra.h \
 auto_qmail.h auto_uids.h date822fmt.h fmtqfn.h
 	./compile qmail-queue.c
 
+eai.o: \
+compile eai.c
+	./compile eai.c
+
 qmail-remote: \
 load qmail-remote.o control.o constmap.o timeoutread.o timeoutwrite.o \
 timeoutconn.o tcpto.o now.o dns.o ip.o ipalloc.o strsalloc.o ipme.o quote.o \
 ndelay.a case.a sig.a open.a lock.a seek.a getln.a stralloc.a alloc.a \
 substdio.a error.a str.a fs.a auto_qmail.o \
-base64.o md5c.o hmac_md5.o \
+base64.o md5c.o hmac_md5.o eai.o \
 dns.lib socket.lib
 	./load qmail-remote control.o constmap.o timeoutread.o \
 	timeoutwrite.o timeoutconn.o tcpto.o now.o dns.o ip.o \
 	tls.o ssl_timeoutio.o -lssl -lcrypto \
 	ipalloc.o strsalloc.o ipme.o quote.o ndelay.a case.a sig.a open.a \
 	lock.a seek.a getln.a stralloc.a alloc.a substdio.a error.a \
-	base64.o md5c.o hmac_md5.o \
+	base64.o md5c.o hmac_md5.o eai.o \
 	str.a fs.a auto_qmail.o `cat dns.lib` `cat socket.lib` -lidn2
 
 qmail-remote.0: \
