+++ TARGETS
@@ -464,3 +464,4 @@ getDomainToken.o
 parse_env.o
 dns_text.o
 conf-policy.temp
+eai.o
